+++ ciri/consensus/tangle/tangle.h
@@ -47,7 +47,7 @@ retcode_t iota_tangle_destroy(tangle_t *const tangle);
  * Transaction operations
  */
 
-retcode_t iota_tangle_transaction_count(tangle_t const *const tangle, size_t *const count);
+retcode_t iota_tangle_transaction_count(tangle_t const *const tangle, uint64_t *const count);
 
 retcode_t iota_tangle_transaction_store(tangle_t const *const tangle, iota_transaction_t const *const tx);
 
@@ -103,7 +103,7 @@ retcode_t iota_tangle_transactions_update_solidity(tangle_t const *const tangle,
                                                    bool const is_solid);
 
 retcode_t iota_tangle_transaction_approvers_count(tangle_t const *const tangle, flex_trit_t const *const hash,
-                                                  size_t *const count);
+                                                  uint64_t *const count);
 
 retcode_t iota_tangle_transactions_metadata_clear(tangle_t const *const tangle);
